splits products plugin
